Add noop.s
noop() is an optimization barrier which disallows aggressive C compilers
to completely remove calls which are considered "pure" (that is, calls
without side-effects). This kind of optimization happens frequently in
LTO builds.

noop() is a function which does nothing, but since it is an external
symbol, it gets disqualified as an LTO target.