Update overflow logic in xatoll/xatoull
This commit changes `jo` to `js` in the logic for detecting overflows,
and also reduces the number of possible exit jumps from two to one.

`jo` is not useful if the previous `lea` has already changed the sign of
the register, which happens most frequently. `js` is the correct way to
detect overflows in this situation.